Migrate persistent data to queryable stores

+++ src/main/java/com/cloudogu/scm/review/pullrequest/dto/TasksDto.java
@@ -26,6 +26,6 @@ import lombok.Setter;
 @NoArgsConstructor
 @AllArgsConstructor
 public class TasksDto {
-  Long todo;
-  Long done;
+  Integer todo;
+  Integer done;
 }
